Moved custom hooks to a more suitable location
