+++ src/components/home/home.container.tsx
@@ -1,6 +1,6 @@
 import { useHistory } from 'react-router';
 import React, { useEffect } from 'react';
-import { useAppDispatch, useAppSelector } from '../../store/hooks';
+import { useAppDispatch, useAppSelector } from '../../hooks/store.hooks';
 import { login, signup } from '../../store/slices/user.slice';
 import routes from '../../routes/route-constants';
 import { SignInState } from '../sign-in/sign-in';
