+++ src/hooks/store.hooks.ts
@@ -1,6 +1,8 @@
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import type { RootState, AppDispatch } from './store';
+import type { AppDispatch, RootState } from '../store/store';
 
-// use the hooks below instead of useDispatch and useSelector
+// --> Use the hooks below instead of useDispatch and useSelector!
+
 export const useAppDispatch = () => useDispatch<AppDispatch>();
+
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
