+++ src/routes/private.route.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Redirect, Route, RouteProps } from 'react-router-dom';
-import { useAppSelector } from '../store/hooks';
+import { useAppSelector } from '../hooks/store.hooks';
 import { selectUser } from '../store/selectors';
 
 interface PrivateRouteProps extends RouteProps {
